+++ ledger-service/http-json-testing/src/main/scala/com/daml/http/Consume.scala
