+++ ledger-service/http-json/BUILD.bazel
@@ -228,6 +228,7 @@ da_scala_test(
         "//ledger-api/rs-grpc-bridge",
         "//ledger-api/testing-utils",
         "//ledger-service/http-json-testing",
+        "//ledger-service/jwt",
         "//ledger/caching",
         "//ledger/ledger-api-auth",
         "//ledger/ledger-api-common",
