+++ ledger-service/http-json/src/failure/scala/http/FailureTests.scala
@@ -3,7 +3,13 @@
 
 package com.daml.http
 
+import akka.http.javadsl.model.ws.PeerClosedConnectionException
 import akka.http.scaladsl.model.{StatusCodes, Uri}
+import akka.stream.scaladsl.Sink
+
+import scala.concurrent.{Future, Promise}
+import scala.util.{Failure, Success}
+import com.daml.http.domain.Offset
 import com.daml.http.json.{JsonError, SprayJson}
 import com.daml.ledger.api.testing.utils.SuiteResourceManagementAroundAll
 import com.daml.timer.RetryStrategy
@@ -25,6 +31,7 @@ final class FailureTests
     with Eventually
     with Inside {
   import HttpServiceTestFixture._
+  import WebsocketTestFixture._
 
   private def headersWithParties(actAs: List[String]) =
     headersWithPartyAuth(actAs, List(), ledgerId().unwrap)
@@ -110,7 +117,6 @@ final class FailureTests
           headersWithParties(List(p.unwrap)))
         _ = status shouldBe StatusCodes.OK
         query = jsObject("""{"templateIds": ["Account:Account"]}""")
-        _ = println("first query")
         (status, output) <- postRequest(
           uri = uri.withPath(Uri.Path("/v1/query")),
           query,
@@ -152,6 +158,81 @@ final class FailureTests
             }
           } yield succeed)
       } yield succeed
+  }
+
+  "/v1/stream/query can reconnect" in withHttpService { (uri, encoder, _, client) =>
+    val query =
+      """[
+          {"templateIds": ["Account:Account"]}
+        ]"""
+
+    val offset = Promise[Offset]()
+
+    def respBefore(accountCid: domain.ContractId): Sink[JsValue, Future[Unit]] = {
+      val dslSyntax = Consume.syntax[JsValue]
+      import dslSyntax._
+      Consume.interpret(
+        for {
+          ContractDelta(Vector((ctId, _)), Vector(), None) <- readOne
+          _ = ctId shouldBe accountCid.unwrap
+          ContractDelta(Vector(), Vector(), Some(liveStartOffset)) <- readOne
+          _ = offset.success(liveStartOffset)
+          _ = proxy.disable()
+          _ <- drain
+        } yield ()
+      )
+    }
+
+    def respAfter(
+        offset: domain.Offset,
+        accountCid: domain.ContractId): Sink[JsValue, Future[Unit]] = {
+      val dslSyntax = Consume.syntax[JsValue]
+      import dslSyntax._
+      Consume.interpret(
+        for {
+          ContractDelta(Vector((ctId, _)), Vector(), Some(newOffset)) <- readOne
+          _ = ctId shouldBe accountCid.unwrap
+          _ = newOffset.unwrap should be > offset.unwrap
+          _ <- drain
+        } yield ()
+      )
+    }
+
+    for {
+      p <- allocateParty(client, "p")
+      (status, r) <- postCreateCommand(
+        accountCreateCommand(p, "abc123"),
+        encoder,
+        uri,
+        headers = headersWithParties(List(p.unwrap)))
+      _ = status shouldBe 'success
+      cid = getContractId(getResult(r))
+      r <- (singleClientQueryStream(
+        jwtForParties(List(p.unwrap), List(), ledgerId().unwrap),
+        uri,
+        query,
+        keepOpen = true) via parseResp runWith respBefore(cid)).transform(x => Success(x))
+      _ = inside(r) {
+        case Failure(e: PeerClosedConnectionException) =>
+          e.closeCode shouldBe 1011
+          e.closeReason shouldBe "internal error"
+      }
+      offset <- offset.future
+      _ = proxy.enable()
+      (status, r) <- postCreateCommand(
+        accountCreateCommand(p, "abc456"),
+        encoder,
+        uri,
+        headers = headersWithParties(List(p.unwrap)))
+      cid = getContractId(getResult(r))
+      _ = status shouldBe 'success
+      _ <- singleClientQueryStream(
+        jwtForParties(List(p.unwrap), List(), ledgerId().unwrap),
+        uri,
+        query,
+        Some(offset)) via parseResp runWith respAfter(offset, cid)
+    } yield succeed
+
   }
 
   protected def jsObject(s: String): JsObject = {
